requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==42.0.8
diff-match-patch==20230430
Django==4.2.15
django-cors-headers==4.3.1
django-import-export==4.1.1
django-maintenance-mode==0.21.1
django-ninja==1.1.0
django-ninja-extra==0.20.7
django-ninja-jwt==5.3.1
django-ninja-passwordreset==1.2.5
django-simple-history==3.7.0
dnspython==2.6.1
email_validator==2.2.0
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
import-export==0.3.1
injector==0.21.0
kombu==5.3.7
ninja-schema==0.13.6
openpyxl==3.1.5
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-fsutil==0.14.1
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
